moved self healing config ui placement
